feat(layout): separate layout
