+++ src/app/(withheader)/layout.tsx
@@ -1,26 +1,11 @@
-import { SharedOptions } from 'msw';
 import { Header } from '~/components/header/Header';
-import { AuthProvider } from '~/feature/auth/AuthProvider';
-import { server } from '~/mocks/server';
 import '~/styles/globals.scss';
 
-if (process.env.NODE_ENV === 'development') {
-  const options: SharedOptions = {
-    onUnhandledRequest: 'bypass'
-  };
-  server.listen(options);
-}
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+export default function WithHeaderLayout({ children }: { children: React.ReactNode }) {
   return (
-    <html lang='ja'>
-      <head />
-      <body>
-        <AuthProvider>
-          <Header />
-          <div>{children}</div>
-        </AuthProvider>
-      </body>
-    </html>
+    <>
+      <Header />
+      <div>{children}</div>
+    </>
   );
 }
